# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "스터디 모집·신청 서비스",
    "description": "\n해커톤 연습용 스터디 모집·신청 서비스 API.\n\n**인증 없음**: 요청자는 `user_id`로만 구분한다.\n- `GET` / `DELETE` → `user_id`를 **쿼리**로 전달\n- `POST` / `PATCH` → `user_id`를 **body**로 전달\n- 방장 전용 API는 `user_id ≠ owner_id`이면 403\n\n에러 응답은 모두 `{\"detail\": \"한국어 메시지\"}` 형식이다. 시간은 KST naive datetime(`YYYY-MM-DDTHH:MM:SS`).\n",
    "version": "1.0.0"
  },
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "사용자 목록",
        "operationId": "list_users_users_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserOut"
                  },
                  "type": "array",
                  "title": "Response List Users Users Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "사용자 생성",
        "description": "닉네임만으로 사용자를 만든다. 닉네임이 중복되면 409.",
        "operationId": "create_user_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "409": {
            "description": "중복/충돌",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}/applications": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "내 신청 목록",
        "description": "사용자가 넣은 모든 신청을 최신순으로 반환한다 (스터디 제목 포함).",
        "operationId": "my_applications_users__user_id__applications_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MyApplicationOut"
                  },
                  "title": "Response My Applications Users  User Id  Applications Get"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplicationStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "APPROVED",
          "REJECTED",
          "CANCELED"
        ],
        "title": "ApplicationStatus"
      },
...
      "ErrorOut": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "examples": [
              "방장만 할 수 있는 작업입니다."
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MyApplicationOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "study_id": {
            "type": "integer",
            "title": "Study Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "study_title": {
            "type": "string",
            "title": "Study Title"
          }
        },
        "type": "object",
        "required": [
          "id",
          "study_id",
          "user_id",
          "message",
          "status",
          "created_at",
          "study_title"
        ],
        "title": "MyApplicationOut"
      },
...
      "UserCreate": {
        "properties": {
          "nickname": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Nickname",
            "examples": [
              "철수"
            ]
          }
        },
        "type": "object",
        "required": [
          "nickname"
        ],
        "title": "UserCreate"
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nickname": {
            "type": "string",
            "title": "Nickname"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nickname"
        ],
        "title": "UserOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}